+++ src/views/e-office/files/notesheet/Notesheet.jsx
@@ -3,8 +3,8 @@ import {EventNote} from "@material-ui/icons";
 
 import Timeline from "../../../../components/Timeline/Timeline.jsx";
 import image from "../../../../assets/img/faces/card-profile1-square.jpg";
-import CreateNoteButton from "./CreateNoteButton";
-import CreateNoteDialog from "../CreateNoteDialog";
+import CreateNoteButton from "./NotesheetCreateButton";
+import CreateNoteDialog from "./NoteCreateDialog";
 import Typography from "@material-ui/core/Typography";
 
 const widgetStories = [